+++ Source~/DangFugBixs.Runtime~/Models/StatModifier.cs
@@ -2,9 +2,9 @@ namespace NhemDangFugBixs.Runtime.Models {
     public struct StatModifier {
         public float Value;
         public ModifierType Type;
-        public object Source;
+        public object? Source;
 
-        public StatModifier(float value, ModifierType type, object source = null) {
+        public StatModifier(float value, ModifierType type, object? source = null) {
             Value = value;
             Type = type;
             Source = source;